Delete filter tools function from MessageManager and move message_manager.py to messages module
